Modernized gem to conform to typical Ruby gem conventions. This includes moving everything under the module RubyManta and adding a Rakefile.
