+++ lib/ruby-manta/version.rb
@@ -1,0 +1,3 @@
+module RubyManta
+  VERSION = '2.0.0'
+end
