--- lib/version.rb
@@ -1,3 +1,0 @@
-class MantaClient
-  LIB_VERSION = '1.2.0'
-end
