+++ ruby-manta.gemspec
@@ -1,26 +1,29 @@
-require File.expand_path('../lib/version', __FILE__)
+require_relative 'lib/ruby-manta/version'
 
 Gem::Specification.new do |s|
   s.name        = 'ruby-manta'
-  s.version     = MantaClient::LIB_VERSION
+  s.version     = RubyManta::VERSION
   s.date        = '2013-10-24'
   s.summary     = "Interface for Joyent's Manta service."
-  s.description = "A simple low-abstraction layer which communicates with Joyent's Manta service." 
+  s.description = "A simple low-abstraction layer which communicates with Joyent's Manta service."
   s.authors     = ['Joyent']
   s.email       = 'marsell@joyent.com'
   s.homepage    = 'http://github.com/joyent/ruby-manta/'
   s.license     = 'MIT'
 
-  s.add_dependency('net-ssh', '>= 2.6.0')
-  s.add_dependency('httpclient', '>= 2.3.0.1')
+  s.add_dependency             'net-ssh', '>= 2.6.0'
+  s.add_dependency             'httpclient', '>= 2.3.0.1'
+
+  s.add_development_dependency 'rake'
+  s.add_development_dependency 'minitest', '~> 5.5.1'
 
   s.files       = ['LICENSE',
                    'README.md',
                    'ruby-manta.gemspec',
                    'example.rb',
-                   'lib/version.rb',
-                   'lib/ruby-manta.rb',
-                   'tests/test_ruby-manta.rb']
+                   'lib/ruby-manta/version.rb',
+                   'lib/ruby-manta/manta_client.rb',
+                   'test/manta_client_test.rb']
 
   s.test_files  = s.files.grep(%r{^test})
   s.require_paths = %w{lib}
